+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog
 ## [1.0.6] - 2019-11-15
 ### Changed
 - 添加 util [wrapt 1.11.2](https://github.com/GrahamDumpleton/wrapt)
+- 添加 test_handler.py 用于方便测试 handlers 方法输出
 
 ## [1.0.5] - 2019-10-06
 ### Fixed